Altered minimum width.
Corrected problem where browsers narrower than about 1260px would not be able to display the contact form appropriately.  Shrinking text was deemed unsuitable as it would become impossibly small to read.  Ideal fix would be to alter page layout for small, or vertical screens (i.e., stacking the form on top of the florist photo), but would result in web page not matching the rubric as instructed.